+++ src/redturtle/prenotazioni/tests/test_week_table_overrides.py
@@ -152,6 +152,189 @@ class TestWeekTableOverridesContextState(unittest.TestCase):
             {"Gate A": self.folder_prenotazioni.week_table},
         )
 
+    @freeze_time("2023-05-14")
+    def test_overrides_in_years_range(self):
+        self.folder_prenotazioni.week_table_overrides = json.dumps(
+            [
+                {
+                    "from_year": "2021",
+                    "from_day": "1",
+                    "from_month": "12",
+                    "to_year": "2024",
+                    "to_day": "1",
+                    "to_month": "3",
+                    "week_table": [
+                        {
+                            "day": "Lunedì",
+                            "morning_start": "1100",
+                            "morning_end": "1200",
+                            "afternoon_start": None,
+                            "afternoon_end": None,
+                        },
+                    ],
+                }
+            ]
+        )
+        now = date.today()
+
+        # if in range, return table overrides
+        self.assertEqual(
+            self.view.get_week_table(date(now.year, 12, 25)),
+            {
+                "Gate A": json.loads(self.folder_prenotazioni.week_table_overrides)[0][
+                    "week_table"
+                ]
+            },
+        )
+
+    @freeze_time("2023-05-14")
+    def test_out_of_years_range_specific_override_not_affects(self):
+        self.folder_prenotazioni.week_table_overrides = json.dumps(
+            [
+                {
+                    "from_year": "2021",
+                    "from_day": "1",
+                    "from_month": "12",
+                    "to_year": "2022",
+                    "to_day": "1",
+                    "to_month": "3",
+                    "week_table": [
+                        {
+                            "day": "Lunedì",
+                            "morning_start": "1100",
+                            "morning_end": "1200",
+                            "afternoon_start": None,
+                            "afternoon_end": None,
+                        },
+                    ],
+                }
+            ]
+        )
+
+        # if out of range, return base table
+        self.assertEqual(
+            self.view.get_week_table(date(2026, 10, 10)),
+            {"Gate A": self.folder_prenotazioni.week_table},
+        )
+
+    @freeze_time("2023-05-14")
+    def test_years_specific_overrides_wins_over_the_generic(self):
+        self.folder_prenotazioni.week_table_overrides = json.dumps(
+            [
+                {
+                    "from_year": "2021",
+                    "from_day": "1",
+                    "from_month": "12",
+                    "to_year": "2024",
+                    "to_day": "1",
+                    "to_month": "3",
+                    "week_table": [
+                        {
+                            "day": "Lunedì",
+                            "morning_start": "1100",
+                            "morning_end": "1200",
+                            "afternoon_start": None,
+                            "afternoon_end": None,
+                        },
+                    ],
+                },
+                {
+                    "from_day": "1",
+                    "from_month": "12",
+                    "to_day": "1",
+                    "to_month": "3",
+                    "week_table": [
+                        {
+                            "day": "Lunedì",
+                            "morning_start": "1100",
+                            "morning_end": "1200",
+                            "afternoon_start": None,
+                            "afternoon_end": None,
+                        },
+                    ],
+                },
+            ],
+        )
+        now = date.today()
+
+        # if in range, return table overrides
+        self.assertEqual(
+            self.view.get_week_table(date(now.year, 12, 25)),
+            {
+                "Gate A": json.loads(self.folder_prenotazioni.week_table_overrides)[0][
+                    "week_table"
+                ]
+            },
+        )
+
+    @freeze_time("2023-05-14")
+    def test_between_the_years_specific_overrides_win_the_last_one(self):
+        self.folder_prenotazioni.week_table_overrides = json.dumps(
+            [
+                {
+                    "from_year": "2021",
+                    "from_day": "1",
+                    "from_month": "12",
+                    "to_year": "2024",
+                    "to_day": "1",
+                    "to_month": "3",
+                    "week_table": [
+                        {
+                            "day": "Lunedì",
+                            "morning_start": "1100",
+                            "morning_end": "1200",
+                            "afternoon_start": None,
+                            "afternoon_end": None,
+                        },
+                    ],
+                },
+                # Expected one
+                {
+                    "from_year": "2021",
+                    "from_day": "1",
+                    "from_month": "12",
+                    "to_year": "2024",
+                    "to_day": "1",
+                    "to_month": "3",
+                    "week_table": [
+                        {
+                            "day": "Lunedì",
+                            "morning_start": "1200",
+                            "morning_end": "1300",
+                            "afternoon_start": None,
+                            "afternoon_end": None,
+                        },
+                    ],
+                },
+                {
+                    "from_day": "1",
+                    "from_month": "12",
+                    "to_day": "1",
+                    "to_month": "3",
+                    "week_table": [
+                        {
+                            "day": "Lunedì",
+                            "morning_start": "1100",
+                            "morning_end": "1200",
+                            "afternoon_start": None,
+                            "afternoon_end": None,
+                        },
+                    ],
+                },
+            ],
+        )
+        now = date.today()
+
+        # if in range, return table overrides
+        self.assertEqual(
+            self.view.get_week_table(date(now.year, 12, 25)),
+            {
+                "Gate A": json.loads(self.folder_prenotazioni.week_table_overrides)[1][
+                    "week_table"
+                ]
+            },
+        )
+
 
 class TestWeekTableOverridesApiValidateDataOnPost(unittest.TestCase):
     layer = REDTURTLE_PRENOTAZIONI_API_FUNCTIONAL_TESTING
@@ -388,3 +571,104 @@ class TestWeekTableOverridesApiValidateDataOnPost(unittest.TestCase):
             "You should set a start time for morning",
             response.json()["message"],
         )
+
+    def test_start_year_greather_than_to_year(self):
+        data = [
+            {
+                "from_year": "2028",
+                "from_day": "1",
+                "from_month": "1",
+                "to_year": "2020",
+                "to_day": "1",
+                "to_month": "2",
+                "week_table": [
+                    {
+                        "day": "Lunedì",
+                        "morning_start": "0700",
+                        "morning_end": "1000",
+                        "afternoon_start": None,
+                        "afternoon_end": None,
+                    },
+                ],
+            }
+        ]
+
+        response = self.api_session.post(
+            self.portal_url,
+            json={
+                "@type": "PrenotazioniFolder",
+                "title": "Example",
+                "daData": json_compatible(date.today()),
+                "week_table_overrides": json.dumps(data),
+                "same_day_booking_disallowed": "yes",
+                "gates": ["Gate A"],
+            },
+        )
+        self.assertEqual(response.status_code, 400)
+
+    def test_only_one_year_field_popolated(self):
+        # only from year
+        data = [
+            {
+                "from_year": "2028",
+                "from_day": "1",
+                "from_month": "1",
+                "to_day": "1",
+                "to_month": "2",
+                "week_table": [
+                    {
+                        "day": "Lunedì",
+                        "morning_start": "0700",
+                        "morning_end": "1000",
+                        "afternoon_start": None,
+                        "afternoon_end": None,
+                    },
+                ],
+            }
+        ]
+
+        response = self.api_session.post(
+            self.portal_url,
+            json={
+                "@type": "PrenotazioniFolder",
+                "title": "Example",
+                "daData": json_compatible(date.today()),
+                "week_table_overrides": json.dumps(data),
+                "same_day_booking_disallowed": "yes",
+                "gates": ["Gate A"],
+            },
+        )
+        self.assertEqual(response.status_code, 400)
+
+        # only to year
+        data = [
+            {
+                "from_day": "1",
+                "from_month": "1",
+                "to_year": "2025",
+                "to_day": "1",
+                "to_month": "2",
+                "week_table": [
+                    {
+                        "day": "Lunedì",
+                        "morning_start": "0700",
+                        "morning_end": "1000",
+                        "afternoon_start": None,
+                        "afternoon_end": None,
+                    },
+                ],
+            }
+        ]
+
+        response = self.api_session.post(
+            self.portal_url,
+            json={
+                "@type": "PrenotazioniFolder",
+                "title": "Example",
+                "daData": json_compatible(date.today()),
+                "week_table_overrides": json.dumps(data),
+                "same_day_booking_disallowed": "yes",
+                "gates": ["Gate A"],
+            },
+        )
+        self.assertEqual(response.status_code, 400)